refactor: util 패키지 생성 및 자동차의 전진여부 담당 클래스 생성.
- 인스턴스변수(상태)없이 기능만을 담당하는 클래스를 '유틸리티 클래스'라고 설정하고 이를 모아놓는 util 패키지를 생성.
- validator와 CarGenerator가 이에 해당됨.
- RandomNumber 클래스에서 랜덤값이 4이상인지 확인만 해주면 랜덤한 값을 value로 저장할 필요가 없음.
- 이를 구현한 메서드를 작성하고 클래스명을 ForwardDecider로 수정함.
- 해당 클래스도 유틸리티 클래스의 성향을 띄기에 해당 패키지로 이동.
