+++ src/main/java/racingcar/util/ForwardDecider.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.util;
 
 import static racingcar.constants.PositionConstant.FORWARD_THRESHOLD;
 import static racingcar.constants.PositionConstant.VALUE_HIGH_BOUND;
@@ -6,18 +6,17 @@ import static racingcar.constants.PositionConstant.VALUE_LOW_BOUND;
 
 import camp.nextstep.edu.missionutils.Randoms;
 
-public class RandomNumber {
-    private final int value;
+public class ForwardDecider {
 
-    public RandomNumber() {
-        value = makeRandomValue();
+    public boolean isForwardPossible() {
+        return checkValueIsOverFour(makeRandomValue());
     }
 
     private int makeRandomValue() {
         return Randoms.pickNumberInRange(VALUE_LOW_BOUND.getValue(), VALUE_HIGH_BOUND.getValue());
     }
 
-    public boolean checkValueIsOverFour() {
-        return value >= FORWARD_THRESHOLD.getValue();
+    private boolean checkValueIsOverFour(int randomValue) {
+        return randomValue >= FORWARD_THRESHOLD.getValue();
     }
 }
