+++ src/test/java/racingcar/util/validator/NameValidatorTest.java
@@ -1,4 +1,4 @@
-package racingcar.validator;
+package racingcar.util.validator;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
@@ -9,11 +9,11 @@ import org.junit.jupiter.params.provider.ValueSource;
 public class NameValidatorTest {
     @DisplayName("이름의 적합성을 판단하는 기능")
     @ParameterizedTest
-    @ValueSource(strings={"eleven","abcdef","qwerdf","111234","1q2w3e4r",""})
-    void validateName_이름의_적합성_판단(String name){
-        NameValidator validator=NameValidator.INSTANCE;
+    @ValueSource(strings = {"eleven", "abcdef", "qwerdf", "111234", "1q2w3e4r", ""})
+    void validateName_이름의_적합성_판단(String name) {
+        NameValidator validator = NameValidator.INSTANCE;
 
-        assertThatThrownBy(()->validator.validateName(name))
+        assertThatThrownBy(() -> validator.validateName(name))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 }
